Trigger Action no longer called on Update()
Removed the binding of starting the game or reloading the level on the Update() function, preventing the binding from occurring every frame. This also has the effect of resolving the issue of needing to press the trigger more than once to start the game when the level has been reloaded.