Parade Implemented, player boundaries placed
Implemented a basic parade with Roaries and football players coming from one side of the field and exiting the other. This is to be played when the player enters roam mode. Also began implementing player boundaries for the upper level of the stadium. Additionally, Dean sits by the stairs of the upper level of the stadium. Placement is not final.